[refactor] move classes
