+++ src/main/java/scs/comp5903/cucumber/EasyCucumber.java
@@ -8,10 +8,10 @@ import scs.comp5903.cucumber.builder.JStepParameterExtractor;
 import scs.comp5903.cucumber.execution.JFeature;
 import scs.comp5903.cucumber.model.exception.EasyCucumberException;
 import scs.comp5903.cucumber.model.exception.ErrorCode;
-import scs.comp5903.cucumber.parser.JStepDefinitionParser;
 import scs.comp5903.cucumber.parser.jfeature.DetailBuilder;
 import scs.comp5903.cucumber.parser.jfeature.JFeatureFileLineByLineParser;
 import scs.comp5903.cucumber.parser.jfeature.JFeatureFileParser;
+import scs.comp5903.cucumber.parser.jstep.JStepDefinitionParser;
 import scs.comp5903.cucumber.util.ReflectionUtil;
 
 import java.nio.file.Path;
