+++ src/test/java/scs/comp5903/cucumber/parser/jstep/JStepDefinitionParserTest.java
@@ -1,4 +1,4 @@
-package scs.comp5903.cucumber.parser;
+package scs.comp5903.cucumber.parser.jstep;
 
 import org.junit.jupiter.api.Test;
 import scs.comp5903.cucumber.model.annotation.step.JStep;
